Fix paper bot: never opened positions (price 400 + whole-unit sizing)
Two bugs stopped the bot buying despite fresh BUY signals:
1. Live-price fetch sent the Binance `symbols` array as JSON WITH spaces, which
   Binance rejects with 400 -> the bot got no prices and skipped everything.
   Now serialised compact (separators=(',',':')).
2. It trusted the signal's whole-unit `suggested_size` (sized to a different
   account), so expensive coins sized to 0 / unaffordable. The bot now sizes
   each position to its OWN equity — an equal-weight slice capped by cash — in
   FRACTIONAL units (quantity -> Numeric). BTC/ETH-priced coins are now tradeable.

Migration d4e5f6a7b8c9 widens bot_positions.quantity to Numeric(24,10).

+++ backend/app/db/models.py
@@ -421,7 +421,8 @@ class BotPosition(Base):
     signal_id: Mapped[int | None] = mapped_column(UBigInt, ForeignKey("signals.id"), nullable=True)
     entry_datetime: Mapped[datetime] = mapped_column(DateTime, nullable=False)
     entry_price: Mapped[Decimal] = mapped_column(Numeric(24, 10), nullable=False)
-    quantity: Mapped[int] = mapped_column(INTEGER, nullable=False)
+    # Fractional — crypto positions are not whole units.
+    quantity: Mapped[Decimal] = mapped_column(Numeric(24, 10), nullable=False)
     stop_price: Mapped[Decimal | None] = mapped_column(Numeric(24, 10), nullable=True)
     horizon_days: Mapped[int] = mapped_column(SMALLINT, nullable=False, server_default=text("10"))
     cost_basis: Mapped[Decimal] = mapped_column(Numeric(24, 10), nullable=False)  # cash spent incl fees
